+++ sdk/cosmosdb/cosmos/src/index.ts
@@ -2,7 +2,6 @@
 // Licensed under the MIT license.
 export { DEFAULT_PARTITION_KEY_PATH } from "./common/partitionKeys";
 export { StatusCodes, StatusCodesType, PartitionKeyRangePropertiesNames } from "./common";
-export { extractPartitionKeys as extractPartitionKey } from "./extractPartitionKey";
 export { setAuthorizationTokenHeaderUsingMasterKey } from "./auth";
 export {
   Operation,
@@ -55,7 +54,9 @@ export {
   PartitionKeyKind,
   PartitionKeyDefinition,
   PartitionKeyDefinitionVersion,
+  PartitionKeyBuilder,
   PermissionMode,
+  PriorityLevel,
   TriggerOperation,
   TriggerType,
   UserDefinedFunctionType,
@@ -66,6 +67,12 @@ export { ContainerRequest } from "./client/Container/ContainerRequest";
 export { Constants, OperationType, ResourceType, HTTPMethod } from "./common";
 export { RetryOptions } from "./retry";
 export * from "./request";
+
+export {
+  DiagnosticNodeInternal,
+  DiagnosticDataValue,
+  DiagnosticNodeType,
+} from "./diagnostics/DiagnosticNodeInternal";
 
 export {
   CosmosHeaders,
@@ -91,12 +98,15 @@ export { ClientContext } from "./ClientContext";
 
 export {
   CosmosDiagnostics,
-  MetadataLookUpDiagnostics,
-  RetryDiagnostics,
   MetadataLookUpDiagnostic,
-  FailedRequestAttemptDiagnostic,
+  MetadataLookUpDiagnostics,
   MetadataLookUpType,
+  RetryDiagnostics,
+  FailedRequestAttemptDiagnostic,
+  GatewayStatistics,
   ClientSideRequestStatistics,
+  ClientConfigDiagnostic,
+  DiagnosticNode,
 } from "./CosmosDiagnostics";
 
 export {
@@ -106,6 +116,7 @@ export {
   ChangeFeedStartFrom,
   FeedRange,
 } from "./client/ChangeFeed";
+export { CosmosDbDiagnosticLevel } from "./diagnostics/CosmosDbDiagnosticLevel";
 
 export { GlobalEndpointManager } from "./globalEndpointManager";
 export { SasTokenPermissionKind } from "./common/constants";