+++ sdk/cosmosdb/cosmos/src/queryExecutionContext/orderByQueryExecutionContext.ts
@@ -1,7 +1,6 @@
 // Copyright (c) Microsoft Corporation.
 // Licensed under the MIT license.
 import { ClientContext } from "../ClientContext";
-import { CosmosDiagnosticContext } from "../CosmosDiagnosticsContext";
 import { PartitionedQueryExecutionInfo } from "../request/ErrorResponse";
 import { FeedOptions } from "../request/FeedOptions";
 import { DocumentProducer } from "./documentProducer";
@@ -34,18 +33,10 @@ export class OrderByQueryExecutionContext
     collectionLink: string,
     query: string | SqlQuerySpec,
     options: FeedOptions,
-    partitionedQueryExecutionInfo: PartitionedQueryExecutionInfo,
-    diagnosticContext: CosmosDiagnosticContext
+    partitionedQueryExecutionInfo: PartitionedQueryExecutionInfo
   ) {
     // Calling on base class constructor
-    super(
-      clientContext,
-      collectionLink,
-      query,
-      options,
-      partitionedQueryExecutionInfo,
-      diagnosticContext
-    );
+    super(clientContext, collectionLink, query, options, partitionedQueryExecutionInfo);
     this.orderByComparator = new OrderByDocumentProducerComparator(this.sortOrders);
   }
   // Instance members are inherited